[#1849] renamed the folder containing the code to release the app
