+++ release/upload-apk/version.properties
@@ -1,0 +1,1 @@
+VERSION=2.0
