--- util/upload-apk/version.properties
@@ -1,1 +1,0 @@
-VERSION=2.0
